requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
colorama==0.4.6
colorlog==6.8.2
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.1
hikari==2.0.0.dev122
hikari-lightbulb==2.3.5
idna==3.6
multidict==6.0.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
yarl==1.9.4
